refactor: ScmService related classes to separate package
